Add useCustomLocator() for non-DOM locators
All DOM-specific parts of the existing locator have been moved into
their own folder, DOMLocator. existing calls to `useLocator()` will
continue to use the DOM locator, but as an opt-in you can provide a
custom locator with the same API as the DOM locator. The types your
custom locator are expected to match can be found at `locator/types.ts`.
